project01/requirements.txt
flask~=3.1.0
flask_sqlalchemy 
psycopg2-binary
opencv-python~=4.11.0.86
flask-socketio~=5.5.1
werkzeug~=3.1.3
numpy~=2.2.4